完善表单

+++ snow-from/src/main/java/com/snow/from/controller/SysFormInstanceController.java
@@ -1,18 +1,25 @@
 package com.snow.from.controller;
 
+import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.collection.CollUtil;
+import cn.hutool.core.util.StrUtil;
 import com.snow.common.annotation.Log;
 import com.snow.common.core.controller.BaseController;
 import com.snow.common.core.domain.AjaxResult;
 import com.snow.common.core.page.TableDataInfo;
+import com.snow.common.core.text.Convert;
 import com.snow.common.enums.BusinessType;
 import com.snow.common.utils.StringUtils;
-import com.snow.common.utils.poi.ExcelUtil;
 import com.snow.framework.util.ShiroUtils;
 import com.snow.from.domain.SysFormDataRecord;
 import com.snow.from.domain.SysFormInstance;
-import com.snow.from.service.ISysFormFieldService;
+import com.snow.from.domain.response.FormDataRecordResponse;
+import com.snow.from.domain.response.FormInstanceResponse;
 import com.snow.from.service.ISysFormInstanceService;
 import com.snow.from.service.impl.SysFormDataRecordServiceImpl;
+import com.snow.system.domain.SysUser;
+import com.snow.system.service.impl.SysUserServiceImpl;
+import org.apache.commons.compress.utils.Lists;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -39,11 +46,13 @@ public class SysFormInstanceController extends BaseController
     @Autowired
     private SysFormDataRecordServiceImpl sysFormDataRecordService;
 
+    @Autowired
+    private SysUserServiceImpl sysUserService;
+
 
     @RequiresPermissions("system:instance:view")
     @GetMapping()
-    public String instance()
-    {
+    public String instance() {
         return prefix + "/instance";
     }
 
@@ -53,30 +62,16 @@ public class SysFormInstanceController extends BaseController
     @RequiresPermissions("system:instance:list")
     @PostMapping("/list")
     @ResponseBody
-    public TableDataInfo list(SysFormInstance sysFormInstance)
-    {
+    public TableDataInfo list(SysFormInstance sysFormInstance) {
         startPage();
         if(!ShiroUtils.isAdmin()){
             Long userId = ShiroUtils.getUserId();
             sysFormInstance.setCreateBy(String.valueOf(userId));
         }
         List<SysFormInstance> list = sysFormInstanceService.selectSysFormInstanceList(sysFormInstance);
-        return getDataTable(list);
+        return getDataTable(warpSysFormInstanceRecord(list));
     }
 
-    /**
-     * 导出单实例列表
-     */
-    @RequiresPermissions("system:instance:export")
-    @Log(title = "单实例", businessType = BusinessType.EXPORT)
-    @PostMapping("/export")
-    @ResponseBody
-    public AjaxResult export(SysFormInstance sysFormInstance)
-    {
-        List<SysFormInstance> list = sysFormInstanceService.selectSysFormInstanceList(sysFormInstance);
-        ExcelUtil<SysFormInstance> util = new ExcelUtil<SysFormInstance>(SysFormInstance.class);
-        return util.exportExcel(list, "instance");
-    }
 
     /**
      * 预览
@@ -100,33 +95,12 @@ public class SysFormInstanceController extends BaseController
         SysFormInstance sysFormInstance = sysFormInstanceService.selectSysFormInstanceById(formId);
         return AjaxResult.success(sysFormInstance.getFromContentHtml());
     }
-    /**
-     * 新增单实例
-     */
-    @GetMapping("/add")
-    public String add()
-    {
-        return prefix + "/add";
-    }
-
-    /**
-     * 新增保存单实例
-     */
-    @RequiresPermissions("system:instance:add")
-    @Log(title = "单实例", businessType = BusinessType.INSERT)
-    @PostMapping("/add")
-    @ResponseBody
-    public AjaxResult addSave(SysFormInstance sysFormInstance)
-    {
-        return toAjax(sysFormInstanceService.insertSysFormInstance(sysFormInstance));
-    }
 
     /**
      * 修改单实例
      */
     @GetMapping("/edit/{id}")
-    public String edit(@PathVariable("id") Long id, ModelMap mmap)
-    {
+    public String edit(@PathVariable("id") Long id, ModelMap mmap) {
         SysFormInstance sysFormInstance = sysFormInstanceService.selectSysFormInstanceById(id);
         mmap.put("sysFormInstance", sysFormInstance);
         return prefix + "/edit";
@@ -139,18 +113,27 @@ public class SysFormInstanceController extends BaseController
     @Log(title = "单实例", businessType = BusinessType.UPDATE)
     @PostMapping("/edit")
     @ResponseBody
-    public AjaxResult editSave(SysFormInstance sysFormInstance)
-    {
-        SysFormInstance sysFormInstanceName = sysFormInstanceService.selectSysFormInstanceByFormCode(sysFormInstance.getFormName());
+    public AjaxResult editSave(SysFormInstance sysFormInstance) {
+        SysFormInstance sysFormInstanceName = sysFormInstanceService.selectSysFormInstanceByFormName(sysFormInstance.getFormName());
+        SysFormInstance sysFormInstanceCode = sysFormInstanceService.selectSysFormInstanceByFormCode(sysFormInstance.getFormName());
         if(StringUtils.isNotNull(sysFormInstanceName)&&!sysFormInstanceName.getFormName().equals(sysFormInstance.getFormName())){
-            return AjaxResult.error(String.format("表单名称:%已存在",sysFormInstance.getFormName()));
+            return AjaxResult.error(StrUtil.format("表单名称{}已存在",sysFormInstance.getFormName()));
+        }
+        if(StringUtils.isNotNull(sysFormInstanceCode)&&!sysFormInstanceCode.getFormCode().equals(sysFormInstance.getFormCode())){
+            return AjaxResult.error(StrUtil.format("表单名称{}已存在",sysFormInstance.getFormCode()));
         }
         return toAjax(sysFormInstanceService.updateSysFormInstance(sysFormInstance));
     }
 
+    /**
+     * 跳转表单记录
+     * @param id 记录id
+     * @param mmap 传给前端的数据
+     * @return 跳转的页面
+     */
     @RequiresPermissions("system:record:view")
     @GetMapping("/formRecord")
-    public String record(@RequestParam Long id,ModelMap mmap)
+    public String formRecord(@RequestParam Long id,ModelMap mmap)
     {
         mmap.put("id",id);
         return prefix + "/record";
@@ -166,9 +149,21 @@ public class SysFormInstanceController extends BaseController
     {
         startPage();
         List<SysFormDataRecord> list = sysFormDataRecordService.selectSysFormDataRecordList(sysFormDataRecord);
-        return getDataTable(list);
+        return getDataTable(warpSysFormDataRecord(list));
     }
 
+    /**
+     * 删除表单记录
+     * @param ids 前端传入的id
+     * @return 返回是否成功
+     */
+    @RequiresPermissions("system:record:remove")
+    @Log(title = "表单记录", businessType = BusinessType.DELETE)
+    @PostMapping( "/removeRecord")
+    @ResponseBody
+    public AjaxResult removeRecord(String ids) {
+        return toAjax(sysFormDataRecordService.deleteSysFormDataRecordByIds(ids));
+    }
 
     /**
      * 删除单实例
@@ -177,8 +172,54 @@ public class SysFormInstanceController extends BaseController
     @Log(title = "单实例", businessType = BusinessType.DELETE)
     @PostMapping( "/remove")
     @ResponseBody
-    public AjaxResult remove(String ids)
-    {
+    public AjaxResult remove(String ids) {
+        SysFormDataRecord sysFormDataRecord = new SysFormDataRecord();
+        sysFormDataRecord.setFormIdList(Convert.toStrArray(ids));
+        List<SysFormDataRecord> sysFormDataRecords = sysFormDataRecordService.selectSysFormDataRecordList(sysFormDataRecord);
+        if(CollUtil.isNotEmpty(sysFormDataRecords)){
+            return AjaxResult.error("该表单已被填写,暂不能删除");
+        }
         return toAjax(sysFormInstanceService.deleteSysFormInstanceByIds(ids));
     }
+
+    /**
+     * 组装参数
+     * @param list 传入需要组装的参数list
+     */
+    private List<FormDataRecordResponse> warpSysFormDataRecord(List<SysFormDataRecord> list ){
+        List<FormDataRecordResponse> formDataRecordResponseList= Lists.newArrayList();
+        if(CollUtil.isEmpty(list)){
+            return formDataRecordResponseList;
+        }
+        list.forEach(t->{
+            SysFormInstance sysFormInstance = sysFormInstanceService.selectSysFormInstanceById(Long.valueOf(t.getFormId()));
+            FormDataRecordResponse formDataRecordResponse =BeanUtil.copyProperties(t,FormDataRecordResponse.class);
+            SysUser sysUser = sysUserService.selectUserById(Long.parseLong(t.getBelongUserId()));
+            formDataRecordResponse.setFormName(sysFormInstance.getFormName());
+            formDataRecordResponse.setBelongUserName(sysUser.getUserName());
+            formDataRecordResponseList.add(formDataRecordResponse);
+        });
+        return formDataRecordResponseList;
+    }
+
+    /**
+     * 组装返回流程实例
+     * @param list 传入list
+     * @return 组装后的数据
+     */
+    private List<FormInstanceResponse> warpSysFormInstanceRecord(List<SysFormInstance> list ){
+        List<FormInstanceResponse> formInstanceResponseList=Lists.newArrayList();
+        if(CollUtil.isEmpty(list)){
+            return formInstanceResponseList;
+        }
+        list.forEach(t->{
+            FormInstanceResponse formInstanceResponse = BeanUtil.copyProperties(t, FormInstanceResponse.class);
+            SysUser sysUser = sysUserService.selectUserById(Long.valueOf(t.getCreateBy()));
+            formInstanceResponse.setCreateByName(sysUser.getUserName());
+            formInstanceResponseList.add(formInstanceResponse);
+        });
+        return formInstanceResponseList;
+    }
+
+
 }
